[ticket/291] Add modules_helper and use for permission checks
B3P-291
